requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cleanup==5.2.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-mailer==2.1
django-sms==0.5.0
lockfile==0.12.2
Pillow==8.4.0
psycopg2==2.9.2
pyserial==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
